+++ RozetkaReactTs/src/components/Header.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+const Header: React.FC = () => {
+
+
+    return (
+        <>
+            <h1>Header</h1>
+        </>
+    );
+};
+
+export default Header;
